refact: 토큰의 타입을 String -> Int로 변환하는 로직을 StringCalculator로 이동
