--- src/test/java/calculator/StringCalculator.kt
@@ -1,10 +1,0 @@
-package calculator
-
-class StringCalculator {
-
-    fun add(text: String?): Int {
-        if (text.isNullOrBlank()) return 0
-        val numbers = Numbers.create(TextParser.parse(text))
-        return numbers.sum()
-    }
-}
