+++ src/test/kotlin/calculator/NumbersTest.kt
@@ -8,10 +8,10 @@ class NumbersTest : DescribeSpec(
     {
         describe("Numbres 클래스의") {
             describe("create 메소드는") {
-                context("String 타입의 숫자가 주어지면") {
+                context("Int 타입의 숫자가 주어지면") {
                     it("Number 인스턴스로 변환해서 상태를 가진다") {
                         assertDoesNotThrow {
-                            Numbers.create(listOf("1", "2", "3"))
+                            Numbers.create(listOf(1, 2, 3))
                         }
                     }
                 }
@@ -19,7 +19,7 @@ class NumbersTest : DescribeSpec(
 
             describe("sum 메소드는") {
                 it("모든 숫자의 합을 반환한다") {
-                    val numbers = Numbers.create(listOf("1", "2", "3"))
+                    val numbers = Numbers.create(listOf(1, 2, 3))
 
                     numbers.sum() shouldBe 6
                 }
