+++ src/test/kotlin/calculator/TextParserTest.kt
